Have one boot config per config options (nodes only)
Instead of having a separate BootConfig/BootGroup for each node not
assigned to a named boot group, this commit adds changes that allow
multiple nodes not in a named boot group to share a
BootConfig/BootGroup. The Add() function will check if a config/group
exists with the kernel/initrd URI and param list first and will use an
existing one, otherwise it will create them and use it.

NOTE: As of this commit, named boot groups are disabled since it needs
to be figured out how to distinguish named groups from XNames (regex?).