Switch out /etcd service endpoint for /data to support multiple backends
BSS used to support only the EtcD backend, so having an /etcd endpoint
made sense. However, with the recent support of a PostgreSQL backend, it
makes more sense to make this more generic.

This endpoint is now changed to /data, and the response JSON now
includes which backend is being used (i.e. "etcd" or "postgres") as well
as the status of the connection to the printed backend.